Update docs and redocs to use dark mode for better readability and theming
