+++ pyproject.toml
@@ -78,7 +78,7 @@ exclude = ["tests*", "deployment*"]
 
 [tool.setuptools.package-data]
 chacc_cli = ["templates/*.template"]
-chacc_api = [".env.sample"]
+chacc_api = [".env.sample", "server/docs/css/*.css"]
 
 [tool.black]
 line-length = 100
